move CachedKernelMixin
